fix: paparazzi tests by moving robolectric tests to different packages
