+++ features/handback/internal/src/test/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/abort/aborted/mobile/AbortedRedirectToMobileWebHolderScreenTest.kt
@@ -1,13 +1,13 @@
-package uk.gov.onelogin.criorchestrator.features.handback.internal.abort.confirm.mobile
+package uk.gov.onelogin.criorchestrator.features.handback.internal.abort.aborted.mobile
 
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertTrue
+import org.junit.Assert
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import uk.gov.onelogin.criorchestrator.features.handback.internal.abort.confirm.mobile.AbortedRedirectToMobileWebHolderScreen
 import uk.gov.onelogin.criorchestrator.features.handback.internal.navigatetomobileweb.FakeWebNavigator
 import uk.gov.onelogin.criorchestrator.features.session.internalapi.domain.REDIRECT_URI
 
@@ -32,12 +32,12 @@ class AbortedRedirectToMobileWebHolderScreenTest {
 
     @Test
     fun `when navigated to, it immediately navigates to mobile web`() {
-        assertEquals(REDIRECT_URI, webNavigator.openUrl)
+        Assert.assertEquals(REDIRECT_URI, webNavigator.openUrl)
     }
 
     @Test
     fun `when navigated to, it immediately runs onFinish`() {
-        assertTrue(hasFinished)
+        Assert.assertTrue(hasFinished)
     }
 
     private fun onFinish() {
