+++ features/handback/internal/src/test/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/abort/confirm/mobile/mobilescreen/ConfirmAbortMobileScreenTest.kt
@@ -1,3 +1,5 @@
+package uk.gov.onelogin.criorchestrator.features.handback.internal.abort.confirm.mobile.mobilescreen
+
 import android.content.Context
 import androidx.compose.ui.test.assertContentDescriptionContains
 import androidx.compose.ui.test.hasText
@@ -29,16 +31,19 @@ import uk.gov.onelogin.criorchestrator.features.session.internalapi.domain.Sessi
 import uk.gov.onelogin.criorchestrator.features.session.internalapi.domain.StubAbortSession
 import uk.gov.onelogin.criorchestrator.features.session.internalapi.domain.createTestInstance
 
+// This test has been moved into it's own package as for some reason, before when this test was
+// in the `mobile` package above, it interferes with Paparazzi screenshot testing
 @RunWith(AndroidJUnit4::class)
 class ConfirmAbortMobileScreenTest {
     @get:Rule
     val composeTestRule = createComposeRule()
     private val context: Context = ApplicationProvider.getApplicationContext()
-    private val continueButton = hasTextStartingWith(context.getString(ConfirmAbortMobileConstants.buttonId))
+    private val continueButton =
+        hasTextStartingWith(context.getString(ConfirmAbortMobileConstants.buttonId))
     private val navController = mock<NavController>()
     private val abortSession = StubAbortSession()
     private val session =
-        Session.createTestInstance(
+        Session.Companion.createTestInstance(
             redirectUri = REDIRECT_URI,
         )
