+++ features/handback/internal/src/test/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/navigation/HandbackNavigationTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.handback.internal
+package uk.gov.onelogin.criorchestrator.features.handback.internal.navigation
 
 import android.content.Context
 import androidx.compose.material3.Button
@@ -26,6 +26,8 @@ import org.mockito.Mockito.mock
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.verifyNoMoreInteractions
 import uk.gov.logging.testdouble.SystemLogger
+import uk.gov.onelogin.criorchestrator.features.handback.internal.HandbackNavGraphProvider
+import uk.gov.onelogin.criorchestrator.features.handback.internal.R
 import uk.gov.onelogin.criorchestrator.features.handback.internal.abort.aborted.desktop.AbortedReturnToDesktopWebViewModelModule
 import uk.gov.onelogin.criorchestrator.features.handback.internal.abort.confirm.desktop.ConfirmAbortDesktopViewModelModule
 import uk.gov.onelogin.criorchestrator.features.handback.internal.abort.confirm.mobile.ConfirmAbortMobileViewModelModule
