+++ features/resume/internal/src/test/kotlin/uk/gov/onelogin/criorchestrator/features/resume/internal/entrypoints/ProveYourIdentityEntryPointsImplTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.resume.internal
+package uk.gov.onelogin.criorchestrator.features.resume.internal.entrypoints
 
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.test.assertIsDisplayed
@@ -13,6 +13,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.kotlin.mock
 import uk.gov.android.ui.theme.m3.GdsTheme
+import uk.gov.onelogin.criorchestrator.features.resume.internal.ProveYourIdentityEntryPointsImpl
 import uk.gov.onelogin.criorchestrator.features.resume.internal.root.ProveYourIdentityViewModel
 import uk.gov.onelogin.criorchestrator.features.resume.internal.root.createTestInstance
 import uk.gov.onelogin.criorchestrator.features.resume.internal.screen.ContinueToProveYourIdentityNavGraphProvider
@@ -24,7 +25,7 @@ class ProveYourIdentityEntryPointsImplTest {
     val composeTestRule = createComposeRule()
 
     private val fakeProveYourIdentityViewModel =
-        ProveYourIdentityViewModel.createTestInstance()
+        ProveYourIdentityViewModel.Companion.createTestInstance()
 
     private val fakeViewModelProviderFactory =
         viewModelFactory {
@@ -52,10 +53,10 @@ class ProveYourIdentityEntryPointsImplTest {
         composeTestRule.setContent {
             GdsTheme {
                 entryPoints.ProveYourIdentityCard(
-                    modifier = Modifier,
+                    modifier = Modifier.Companion,
                 )
             }
         }
-        composeTestRule.onNodeWithTag(ProveYourIdentityEntryPointsImpl.TEST_TAG).assertIsDisplayed()
+        composeTestRule.onNodeWithTag(ProveYourIdentityEntryPointsImpl.Companion.TEST_TAG).assertIsDisplayed()
     }
 }
